+++ text/src/main/java/eu/interedition/text/query/RangeLengthCriterion.java
@@ -4,13 +4,13 @@ package eu.interedition.text.query;
  * @author <a href="http://gregor.middell.net/" title="Homepage">Gregor Middell</a>
  */
 public class RangeLengthCriterion implements Criterion {
-  private final int length;
+  private final long length;
 
-  public RangeLengthCriterion(int length) {
+  public RangeLengthCriterion(long length) {
     this.length = length;
   }
 
-  public int getLength() {
+  public long getLength() {
     return length;
   }
 }
